Удаление и получение пользователей

+++ api/KpoApi/KpoApi.Domain/Entities/User.cs
@@ -8,7 +8,7 @@ public class User
     public Guid UserUuid { get; init; }
     public string Login { get; init; }
     public string? Password { get; set; }
-    public int? PhoneNumber { get; init; }
+    public long? PhoneNumber { get; init; }
     public string FullName { get; init; }
     public Guid OrganizationUuid { get; init; }
     public string Appointment { get; init; }
